Add template and remove devtool tracking
Added a template html file to the src folder and changed the settings
in webpack config so the output would use the template. Altered code in
page.js file so that the elements would append to the content section.
Added any runtime files to the gitignore file and untracked the runtime
files because they are only necessary for development