Add elements and modify README
Added a pizza img to the project and altered the page.js file in order
to append it to the page. Also added logic to the page.js file in order
to create and append a title in the header for the restaurant name
and a testimonial in the main section. Added styling to the img and the
page overall to begin making it look nicer