Add main page and styling
Added settings to webpack to alter the filepath and naming scheme of the
images so they would appear as the same name as their source rather than
being a random hash name. Added logic to append various elements for the
homepage and styled the page to make it more presentable. Got rid of
debugging function in index.js.