Add page creation modules and photos
Added images to the project in order to add them to the selection boxes
in the main section. Adjusted the README file to properly credit the
photographers. Added button functionality to index.js so that when the
menu button was pressed the menu tab would load. Added menu module that
renders all of the menu content on the page when called. Added home
module that re-renders the home page when called. Adjusted styling of
the menu page so that the photos would look better in their cards.