Add contact page module
Added module for rendering the contact page with all of the placeholder
info. Adjusted other modules to have the contact module loader button
functionality so that the contact page would render whenever the button
was pressed on any page. Adjusted styling of contact info in order to
make it more readable.